refactor: handler함수 분리 (DASOMFE-34)
